requirements.txt
flask
requests
slackclient
bs4